Insignificant changes in API: int->long, new action types in preparation for snapshot
git-svn-id: https://devel.codership.com/svn/galera/trunk@458 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ gcs/src/gcs_conf.c
@@ -6,16 +6,19 @@
 
 /* Logging options */
 #include <galerautils.h>
-int gcs_conf_set_log_file     (FILE *file)
+
+#include "gcs.h"
+
+long gcs_conf_set_log_file     (FILE *file)
 { return gu_conf_set_log_file (file); }
-int gcs_conf_set_log_callback (void (*logger) (int, const char*))
+long gcs_conf_set_log_callback (void (*logger) (int, const char*))
 { return gu_conf_set_log_callback (logger); }
-int gcs_conf_self_tstamp_on   ()
+long gcs_conf_self_tstamp_on   ()
 { return gu_conf_self_tstamp_on (); }
-int gcs_conf_self_tstamp_off  ()
+long gcs_conf_self_tstamp_off  ()
 { return gu_conf_self_tstamp_off (); }
-int gcs_conf_debug_on  ()
+long gcs_conf_debug_on  ()
 { return gu_conf_debug_on (); }
-int gcs_conf_debug_off ()
+long gcs_conf_debug_off ()
 { return gu_conf_debug_off (); }
